Remove TimerArgs
This simplifies the code, and means that `next` is set at the point the
`Timer` registration is requested by the code that needs it. This code
should be responsible for setting next as it will be responding to e.g.
the need to perform some action 1 second after the code itself runs.