Improve compression module interface
The compression code was previously copied from a crate I wrote which
exposed an interface allowing different huffman tables and terminal
codes to be used with it.

In this server the table and terminal code will always be static, so
this commit updates the compression interface to encapsulate them inside
the module.